# backend/my-blog-hono/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "QBlog Hono API",
		"description": "RESTful API for QBlog - Blog and Portfolio Management System",
		"version": "1.0.0",
		"contact": {
			"name": "QBlog Support",
			"url": "https://qblog.dev"
		},
		"license": {
			"name": "MIT"
		}
	},
	"servers": [
		{
			"url": "http://localhost:8787",
			"description": "Local Development Server"
		},
		{
			"url": "https://api.qblog.dev",
			"description": "Production Server"
		}
	],
...
	"paths": {
...
		"/login": {
			"post": {
				"tags": ["Authentication"],
				"summary": "Admin Login",
				"description": "Authenticate admin user and create session",
				"operationId": "login",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Login successful",
						"headers": {
							"Set-Cookie": {
								"schema": {
									"type": "string",
									"example": "session_id=abc123; Path=/; HttpOnly; Secure; SameSite=Strict"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse"
								}
							}
						}
					},
					"401": {
						"description": "Invalid credentials",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/logout": {
			"post": {
				"tags": ["Authentication"],
				"summary": "Admin Logout",
				"description": "Destroy admin session",
				"operationId": "logout",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Logout successful",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse"
								}
							}
						}
					}
				}
			}
		},
		"/auth/check": {
			"get": {
				"tags": ["Authentication"],
				"summary": "Check Auth Status",
				"description": "Verify if user is authenticated",
				"operationId": "checkAuth",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Auth status retrieved",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse"
								}
							}
						}
					}
				}
			}
		},
		"/api/articles": {
			"get": {
				"tags": ["Articles"],
				"summary": "List Published Articles",
				"description": "Get paginated list of published articles",
				"operationId": "listArticles",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"description": "Page number (1-indexed)",
						"schema": {
							"type": "integer",
							"default": 1,
							"minimum": 1
						}
					},
					{
						"name": "per_page",
						"in": "query",
						"description": "Items per page (1-24)",
						"schema": {
							"type": "integer",
							"default": 6,
							"minimum": 1,
							"maximum": 24
						}
					}
				],
				"responses": {
					"200": {
						"description": "Articles retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ArticlesListResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Articles"],
				"summary": "Create Article",
				"description": "Create a new article (Admin only)",
				"operationId": "createArticle",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateArticleRequest"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Article created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ArticleResponse"
								}
							}
						}
					},
					"400": {
						"description": "Missing required fields",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/api/auto-save": {
			"post": {
				"tags": ["Articles"],
				"summary": "Auto-save Article Draft",
				"description": "Auto-save or create article draft (Admin only)",
				"operationId": "autoSaveArticle",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AutoSaveArticleRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Article auto-saved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AutoSaveResponse"
								}
							}
						}
					},
					"400": {
						"description": "Missing title",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/blog/{slug}": {
			"get": {
				"tags": ["Articles"],
				"summary": "Get Article by Slug",
				"description": "Retrieve a single published article by slug. Increments view count.",
				"operationId": "getArticleBySlug",
				"parameters": [
					{
						"name": "slug",
						"in": "path",
						"required": true,
						"description": "Article slug (URL-friendly identifier)",
						"schema": {
							"type": "string",
							"example": "my-first-article"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Article retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ArticleDetailResponse"
								}
							}
						}
					},
					"404": {
						"description": "Article not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["Articles"],
				"summary": "Update Article",
				"description": "Update an article by slug (Admin only)",
				"operationId": "updateArticle",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"parameters": [
					{
						"name": "slug",
						"in": "path",
						"required": true,
						"description": "Article slug",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateArticleRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Article updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ArticleResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Article not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Articles"],
				"summary": "Delete Article",
				"description": "Delete an article by slug (Admin only)",
				"operationId": "deleteArticle",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"parameters": [
					{
						"name": "slug",
						"in": "path",
						"required": true,
						"description": "Article slug",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Article deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Article not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"cookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "session_id"
			}
		},
		"schemas": {
...
			"SuccessResponse": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string",
						"example": "success"
					},
					"data": {
						"type": "object"
					}
				}
			},
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string",
						"example": "error"
					},
					"error": {
						"type": "string",
						"example": "Error message"
					}
				}
			},
			"LoginRequest": {
				"type": "object",
				"required": ["username", "password"],
				"properties": {
					"username": {
						"type": "string",
						"example": "olloyor"
					},
					"password": {
						"type": "string",
						"example": "your_password"
					}
				}
			},
			"CreateArticleRequest": {
				"type": "object",
				"required": ["title", "content"],
				"properties": {
					"title": {
						"type": "string",
						"example": "My First Article"
					},
					"content": {
						"type": "string",
						"example": "<p>Article content in HTML</p>"
					},
					"is_published": {
						"type": "boolean",
						"default": false,
						"example": false
					}
				}
			},
			"UpdateArticleRequest": {
				"type": "object",
				"properties": {
					"title": {
						"type": "string",
						"example": "Updated Title"
					},
					"content": {
						"type": "string",
						"example": "<p>Updated content</p>"
					},
					"is_published": {
						"type": "boolean",
						"example": true
					}
				}
			},
			"AutoSaveArticleRequest": {
				"type": "object",
				"required": ["title"],
				"properties": {
					"title": {
						"type": "string",
						"example": "Draft Title"
					},
					"content": {
						"type": "string",
						"example": "Draft content"
					},
					"slug": {
						"type": "string",
						"description": "Optional - if provided, updates existing article; if omitted, creates new draft",
						"example": "my-article"
					}
				}
			},
			"Article": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"example": 1
					},
					"slug": {
						"type": "string",
						"example": "my-first-article"
					},
					"title": {
						"type": "string",
						"example": "My First Article"
					},
					"summary": {
						"type": "string",
						"example": "This is a summary of the article..."
					},
					"published_on": {
						"type": "string",
						"example": "28 October 2025"
					},
					"reading_time": {
						"type": "string",
						"example": "5 min read"
					}
				}
			},
			"ArticleDetail": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"slug": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"is_published": {
						"type": "boolean"
					},
					"published_on": {
						"type": "string"
					},
					"reading_time": {
						"type": "string"
					},
					"viewCount": {
						"type": "integer",
						"example": 42
					}
				}
			},
			"ArticlesListResponse": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string",
						"example": "success"
					},
					"data": {
						"type": "object",
						"properties": {
							"articles": {
								"type": "array",
								"items": {
									"$ref": "#/components/schemas/Article"
								}
							},
							"has_more": {
								"type": "boolean",
								"example": true
							},
							"total": {
								"type": "integer",
								"example": 49
							}
						}
					}
				}
			},
			"ArticleResponse": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string",
						"example": "success"
					},
					"data": {
						"$ref": "#/components/schemas/ArticleDetail"
					}
				}
			},
			"ArticleDetailResponse": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string",
						"example": "success"
					},
					"data": {
						"$ref": "#/components/schemas/ArticleDetail"
					}
				}
			},
			"AutoSaveResponse": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string",
						"example": "success"
					},
					"data": {
						"type": "object",
						"properties": {
							"slug": {
								"type": "string",
								"example": "my-article"
							}
						}
					}
				}
			},
...
			"DeleteResponse": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string",
						"example": "success"
					},
					"data": {
						"type": "object",
						"properties": {
							"id": {
								"type": "string",
								"example": "slug_or_id"
							}
						}
					}
				}
			}
		}
	}
}